Add background color feature
Functions were made to enable the background coloring of each pixel
in the grid based on the users selection. Also, the user choice
is highlighted by a box shadow. The box shadow is reset with each
new selection. The background color is parsed from the clicked div
to the function responsible for coloring the background.