Add dynamic background and pixel coloring func.
Bug concerning missing background coloring when grid resolution
was changed was fixed. It was a problem with the gridPixels variable
that should be re-defined by createGrid().
Also, pen-coloring feature for each pixel was implemented as part of
createGrid() function with the penColor variable being defined by user
choice via new function.