Update new color picker choice for pen and misc.
A new function was implemented to keep updating the pen color when
using the color picker. Previously the color value for the pen
would remain the same even though a new color was chosen. An
"input" event was used to trigger the updating function.

Also, the clear button was completed by calling the existing function
updateBackground() when pressed. This clears the sketch board but
keeps the resolution and color choices.