Add new variable to store if rainbow pen is select
DOM-related variable declarations have been changed from
"let" to "const.

A new variable has been created to store is the rainbow pen is
selected or not. This will be used to determine when the random color
function must be envoked by mousehover mouseevent to paint each pixel
a new random color if this variable goes high.