Add random pixel color feature
A feature was added which enables the user to apply a random
rgba color to each pixel if this feature is chosen by the user.
The random coloring will apply both when a grid pixel is clicked and
with subsequent mouseover events.

Also, some housekeeping and re-naming of a few variables and function
names was applied to the code to finalize.